requirements.txt
tensorflow

notebook
jupyterlab==3.6.6

numpy
pandas 
scikit-learn
scipy

matplotlib
seaborn

python-box
pyYAML
types-PyYAML
tqdm
ensure
joblib

Flask
Flask-Cors

-e .